ADD: ZEUIControl added. ADD: ZEUIRenderer added. BUG: DirectInput InputMap->InputActionCount crash bug fixed. ADD: ZEFontResource added. ADD: AddReferance has been added to ZEResource. It will be used on many Setter and Getters and will track shared and non shared resources. CHG: Now unshared resources have referance count. CHG: Resource release mechanism changed. Referance count added for non shared resources.
svn path=/trunk/; revision=95

+++ Source/Core/ConsoleWindow.cpp
@@ -147,7 +147,7 @@ bool ZEConsoleWindow::Initialize()
 
 	ShowWindow((HWND)Handle, SW_HIDE);
 
-	for (int I = 0; I < zeConsole->OutputHistory.GetCount(); I++)
+	for (size_t I = 0; I < zeConsole->OutputHistory.GetCount(); I++)
 		Output(zeConsole->OutputHistory[I]);
 
 	return true;
